Addressed the review comment of splitting one main file into multiple files, event handling into one, rpc into one, polling into one
